Enable CONFIG_FS_VERITY
We'd like to experiment with this for Fedora CoreOS, see:
https://github.com/coreos/coreos-assembler/pull/876

Because it operates at the file level, fs-verity is significantly
more flexible than dm-verity and can be used in many more ways.

While there will be a lot of follow-on work required to make
it truly secure, as soon as XFS gains support I think we'd
turn it on unconditionally.

Since making use of this requires a ext4/f2fs filesystem explicitly
created or enabled, I believe it's quite safe to turn on.