Force kernel to generate DWARF4 debuginfo
Currently, some tools such as crash utility do not support the DWARF5.
But unfortunately, the DWARF5 is enabled on Fedora34/35 by default,
which caused the failure of analyzing vmcore.

In view of this, force kernel to generate DWARF4 debuginfo via the
CONFIG_DEBUG_INFO_DWARF4 kernel option in order to make some tools(crash
utility) happy.

Signed-off-by: Lianbo Jiang <lijiang@redhat.com>